+++ dat/util.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/mgutz/str"
 
-	"gopkg.in/mgutz/dat.v2/common"
+	"gopkg.in/mgutz/dat.v3/common"
 )
 
 // NameMapping is the routine to use when mapping column names to struct properties