+++ replace.sh
@@ -1,3 +1,3 @@
-find . -name "*.go" -not -path "./node_modules/*" -print | xargs sed -i 's|"gopkg.in/mgutz/dat.v2"|"gopkg.in/mgutz/dat.v2/dat"|g'
+find . -name "*.go" -not -path "./node_modules/*" -print | xargs sed -i 's|"gopkg.in/mgutz/dat.v3"|"gopkg.in/mgutz/dat.v3/dat"|g'