+++ sqlx-runner/benchmark_insert_transaction_test.go
@@ -3,7 +3,7 @@ package runner
 import (
 	"testing"
 
-	"gopkg.in/mgutz/dat.v2/dat"
+	"gopkg.in/mgutz/dat.v3/dat"
 )
 
 // These benchmarks compare the total cost of interpolating the SQL then