+++ sqlx-runner/benchmark_varying_length_test.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"testing"
 
-	"gopkg.in/mgutz/dat.v2/dat"
+	"gopkg.in/mgutz/dat.v3/dat"
 )
 
 // These benchmarks compare the total cost of interpolating the SQL then