+++ sqlx-runner/cache_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"gopkg.in/mgutz/dat.v2/dat"
+	"gopkg.in/mgutz/dat.v3/dat"
 	"gopkg.in/stretchr/testify.v1/assert"
 
 	"github.com/mgutz/jo/v1"