+++ sqlx-runner/init.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/cenkalti/backoff"
 	"github.com/mgutz/logxi/v1"
-	"gopkg.in/mgutz/dat.v2/dat"
-	"gopkg.in/mgutz/dat.v2/kvs"
-	"gopkg.in/mgutz/dat.v2/postgres"
+	"gopkg.in/mgutz/dat.v3/dat"
+	"gopkg.in/mgutz/dat.v3/kvs"
+	"gopkg.in/mgutz/dat.v3/postgres"
 )
 
 // Logger is the internal logger interface