+++ sqlx-runner/insert_exec_test.go
@@ -6,9 +6,9 @@ import (
 	"strings"
 	"testing"
 
-	"gopkg.in/mgutz/dat.v2/common"
-	"gopkg.in/mgutz/dat.v2/dat"
-	"gopkg.in/mgutz/dat.v2/postgres"
+	"gopkg.in/mgutz/dat.v3/common"
+	"gopkg.in/mgutz/dat.v3/dat"
+	"gopkg.in/mgutz/dat.v3/postgres"
 	"gopkg.in/stretchr/testify.v1/assert"
 )